Safari-609.1.11.1/LayoutTests/fast/frames/frame-focus-blurs-active-element-expected.txt

CONSOLE MESSAGE: line 17: PASS: blur event was fired on the active element when the iframe was focused.
 
